# api/client.v1.swagger.yml
openapi: 3.0.3
info:
  title: Bank Support Chat Client API
  description: |
    API для сервиса чатов. 
    Все запросы осуществляются методом POST.
    Все ответы имеют формат {"data": {...}} или {"error": {...}}.
    
    Примечание: API доступно через базовый URL: http://localhost:8080
    Все пути в спецификации указаны относительно этого URL.
  version: v1

servers:
  - url: http://localhost:8080/v1
    description: Development server

paths:
  /getHistory:
    post:
      operationId: PostGetHistory
      description: Get chat history.
      parameters:
        - $ref: "#/components/parameters/XRequestIDHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GetHistoryRequest"
      responses:
        '200':
          description: Messages list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetHistoryResponse"

security:
  - bearerAuth: [ ]

components:
  # Компоненты безопасности
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    # Заголовок запроса
    XRequestIDHeader:
      in: header
      name: X-Request-ID
      description: Unique request identifier
      schema:
        type: string
        format: uuid
        x-go-type: types.RequestID
        x-go-type-import:
          path: "github.com/FischukSergey/chat-service/internal/types"
      required: true

  schemas:
    # Common.

    Error:
      required: [ message, code ]
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
        details:
          type: string

    ErrorCode:
      type: integer
      description: contains HTTP error codes and specific business logic error codes (the last must be >= 1000).
      # TODO: Конкретные коды появятся в следующем уроке.
      minimum: 400

    # /getHistory

    GetHistoryRequest:
      type: object
      properties:
        pageSize:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          nullable: true
        cursor:
          type: string
          nullable: true
          description: Курсор для пагинации

    GetHistoryResponse:
      properties:
        data:
          $ref: "#/components/schemas/MessagesPage"
        error:
          $ref: "#/components/schemas/Error"

    MessagesPage:
      type: object
      required: [ messages ] 
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        nextCursor:
          type: string
          nullable: true
          description: |
            Курсор для следующей страницы.
            Если нет следующей страницы, то не возвращается.
            Если нет курсора, то возвращается пустая строка.
        prevCursor:
          type: string
          nullable: true
          description: |
            Курсор для предыдущей страницы.
            Если нет предыдущей страницы, то не возвращается.
            Если нет курсора, то возвращается пустая строка.

    Message:
      type: object
      required: [ id, body, createdAt, isReceived, isBlocked, isService ] 
      properties:
        id:
          type: string
          format: uuid
          x-go-type: types.MessageID
          x-go-type-import:
            path: "github.com/FischukSergey/chat-service/internal/types"
        authorId:
          type: string
          format: uuid
          x-go-type: types.UserID
          x-go-type-import:
            path: "github.com/FischukSergey/chat-service/internal/types"
        body:
          type: string
        createdAt:
          type: string
          format: date-time
        isReceived:
          type: boolean
        isBlocked:
          type: boolean
        isService:
          type: boolean